+++ riana/gui/protein_tab.py
@@ -245,6 +245,7 @@ class ProteinTab(QWidget):
         self.model = DataFrameTableModel()
         self.table.setModel(self.model)
         self.table.setSelectionBehavior(QTableView.SelectionBehavior.SelectRows)
+        self.table.setSortingEnabled(True)
         self.table.selectionModel().currentRowChanged.connect(self._on_row_changed)
         results_split.addWidget(self.table)